+++ linux/ecosystem/torrust-tracker/docker-entrypoint.sh
@@ -2,8 +2,8 @@
 set -euo pipefail
 
 # Setup default Opts
-: ${RACKER_PORT:=1337}
-: ${RACKER_ADMIN:=1488}
+: ${TORRUST_PORT:=1337}
+: ${TORRUST_ADMIN:=1488}
 
 echo "[torrust-tracker] Starting up"
 cd /app